move api from low_level_rs to pipeline
