+++ springql-core/src/pipeline.rs
@@ -17,21 +17,129 @@ pub(crate) mod stream_model;
 pub(crate) mod test_support;
 
 use anyhow::anyhow;
+use std::sync::Once;
+use std::thread;
+use std::time::Duration;
 use std::{collections::HashSet, sync::Arc};
 
-use crate::error::{Result, SpringError};
-use crate::stream_engine::{SinkRow, SqlValue};
-use crate::SpringValue;
-
-use self::{
-    name::StreamName, pipeline_graph::PipelineGraph, pipeline_version::PipelineVersion,
-    pump_model::PumpModel, sink_writer_model::SinkWriterModel,
-    source_reader_model::SourceReaderModel, stream_model::StreamModel,
+use crate::pipeline::name::QueueName;
+use crate::{
+    api::low_level_rs::EngineMutex,
+    error::{Result, SpringError},
+    low_level_rs::SpringConfig,
+    pipeline::{
+        name::StreamName, pipeline_graph::PipelineGraph, pipeline_version::PipelineVersion,
+        pump_model::PumpModel, sink_writer_model::SinkWriterModel,
+        source_reader_model::SourceReaderModel, stream_model::StreamModel,
+    },
+    sql_processor::SqlProcessor,
+    stream_engine::{command::Command, SinkRow, SqlValue},
+    SpringValue,
 };
 
-pub use crate::api::low_level_rs::{
-    spring_command, spring_open, spring_pop, spring_pop_non_blocking, SpringPipeline,
-};
+fn setup_logger() {
+    static INIT: Once = Once::new();
+
+    INIT.call_once(|| {
+        let _ = env_logger::builder()
+            .is_test(false) // To enable color. Logs are not captured by test framework.
+            .try_init();
+        log_panics::init();
+    });
+
+    log::info!("setup_logger(): done");
+}
+
+/// Connection object.
+///
+/// 1 stream pipeline has only 1 connection.
+/// In other words, the lifecycle of SpringConnection and internal stream pipeline are the same.
+#[derive(Debug)]
+pub struct SpringPipeline {
+    engine: EngineMutex,
+    sql_processor: SqlProcessor,
+}
+
+/// Creates and open an in-process stream pipeline.
+pub fn spring_open(config: &SpringConfig) -> Result<SpringPipeline> {
+    setup_logger();
+
+    let engine = EngineMutex::new(config);
+    let sql_processor = SqlProcessor::default();
+
+    Ok(SpringPipeline {
+        engine,
+        sql_processor,
+    })
+}
+
+/// Execute commands (DDL).
+///
+/// # Failure
+///
+/// - [SpringError::Sql](crate::error::SpringError::Sql) when:
+///   - Invalid SQL syntax.
+///   - Refers to undefined objects (streams, pumps, etc)
+///   - Other semantic errors.
+/// - [SpringError::InvalidOption](crate::error::SpringError::Sql) when:
+///   - `OPTIONS` in `CREATE` statement includes invalid key or value.
+pub fn spring_command(pipeline: &SpringPipeline, sql: &str) -> Result<()> {
+    let mut engine = pipeline.engine.get()?;
+
+    let command = pipeline
+        .sql_processor
+        .compile(sql, engine.current_pipeline())?;
+
+    match command {
+        Command::AlterPipeline(c) => engine.alter_pipeline(c),
+    }
+}
+
+/// Pop a row from an in memory queue. This is a blocking function.
+///
+/// **Do not call this function from threads.**
+/// If you need to pop from multiple in-memory queues using threads, use `spring_pop_non_blocking()`.
+/// See: <https://github.com/SpringQL/SpringQL/issues/125>
+///
+/// # Failure
+///
+/// - [SpringError::Unavailable](crate::error::SpringError::Unavailable) when:
+///   - queue named `queue` does not exist.
+pub fn spring_pop(pipeline: &SpringPipeline, queue: &str) -> Result<SpringRow> {
+    const SLEEP_MSECS: u64 = 10;
+
+    let mut engine = pipeline.engine.get()?;
+
+    loop {
+        if let Some(sink_row) =
+            engine.pop_in_memory_queue_non_blocking(QueueName::new(queue.to_string()))?
+        {
+            return Ok(SpringRow::from(sink_row));
+        } else {
+            thread::sleep(Duration::from_millis(SLEEP_MSECS));
+        }
+    }
+}
+
+/// Pop a row from an in memory queue. This is a non-blocking function.
+///
+/// # Returns
+///
+/// - `Ok(Some)` when at least a row is in the queue.
+/// - `None` when no row is in the queue.
+///
+/// # Failure
+///
+/// - [SpringError::Unavailable](crate::error::SpringError::Unavailable) when:
+///   - queue named `queue` does not exist.
+pub fn spring_pop_non_blocking(
+    pipeline: &SpringPipeline,
+    queue: &str,
+) -> Result<Option<SpringRow>> {
+    let mut engine = pipeline.engine.get()?;
+    let sink_row = engine.pop_in_memory_queue_non_blocking(QueueName::new(queue.to_string()))?;
+    Ok(sink_row.map(SpringRow::from))
+}
 
 /// Get an integer column.
 ///
